+++ src/compile/chunks/PhpValueInterface.php
@@ -1,9 +1,11 @@
 <?php
 namespace VovanVE\HtmlTemplate\compile\chunks;
 
+use VovanVE\HtmlTemplate\compile\CompileScope;
+
 interface PhpValueInterface
 {
-    public function getPhpCode(): string;
+    public function getPhpCode(CompileScope $scope): string;
 
     public function isConstant(): bool;