+++ tests/unit/compile/cases/tags/expr/23.logic-or-const.t
@@ -1,0 +1,17 @@
+<a>{ true || true || 'fail' }</a>{'\n'}
+<b>{ true || false || 'fail' }</b>{'\n'}
+<c>{ false || true || 'fail' }</c>{'\n'}
+<d>{ false || false || 'ok' }</d>{'\n'}
+<e>{ false || (false || 'ok') }</e>
+---- CODE ----
+'<a></a>
+<b></b>
+<c></c>
+<d>ok</d>
+<e>ok</e>'
+---- RESULT ----
+<a></a>
+<b></b>
+<c></c>
+<d>ok</d>
+<e>ok</e>