+++ tests/unit/compile/cases/tags/expr/31.ternary-var.t
@@ -1,0 +1,5 @@
+<a>{ lorem ? ipsum : dolor }</a>
+---- CODE ----
+($runtime::createElement('a',[],[($runtime::htmlEncode(((($runtime->param('lorem')))?(($runtime->param('ipsum'))):(($runtime->param('dolor'))))))]))
+---- RESULT ----
+<a>[value of &amp;ipsum]</a>